requirements.txt
selenium==4.10.0
pillow==9.5.0
pytesseract==0.3.10
schedule==1.2.0